+++ Tests/Unit/Stubs/Microsoft365.psm1
@@ -73195,6 +73195,10 @@ function New-CsTeamsCallingPolicy
         $RealTimeText,
 
         [Parameter()]
+        [System.String]
+        $ReportCall,
+
+        [Parameter()]
         [System.Boolean]
         $ShowTeamsCallsInCallLog,
 
@@ -73468,6 +73472,10 @@ function New-CsTeamsEventsPolicy
 
         [Parameter()]
         [System.String]
+        $BackroomChat,
+
+        [Parameter()]
+        [System.String]
         $Identity,
 
         [Parameter()]
@@ -73513,6 +73521,10 @@ function New-CsTeamsEventsPolicy
         [Parameter()]
         [System.String]
         $TownhallEventAttendeeAccess,
+
+        [Parameter()]
+        [System.String]
+        $TownhallMaxResolution,
 
         [Parameter()]
         [System.String]
@@ -73631,6 +73643,10 @@ function New-CsTeamsMeetingPolicy
         $AllowMeetingCoach,
 
         [Parameter()]
+        [System.Boolean]
+        $AllowMultipleScreenshare,
+
+        [Parameter()]
         [System.String]
         $AttendeeIdentityMasking,
 
@@ -73653,6 +73669,10 @@ function New-CsTeamsMeetingPolicy
         [Parameter()]
         [System.Boolean]
         $AllowAnonymousUsersToDialOut,
+
+        [Parameter()]
+        [System.String]
+        $BackroomChat,
 
         [Parameter()]
         [System.String]
@@ -73773,6 +73793,10 @@ function New-CsTeamsMeetingPolicy
         [Parameter()]
         [System.Boolean]
         $AllowWatermarkForScreenSharing,
+
+        [Parameter()]
+        [System.String]
+        $PasscodeComplexity,
 
         [Parameter()]
         [System.String]
@@ -73961,6 +73985,10 @@ function New-CsTeamsMeetingPolicy
         [Parameter()]
         [System.Boolean]
         $AllowPSTNUsersToBypassLobby,
+
+        [Parameter()]
+        [System.String]
+        $SetRecordingAndTranscriptOwnership,
 
         [Parameter()]
         [System.String]
@@ -75026,6 +75054,10 @@ function Set-CsTeamsAIPolicy
 
         [Parameter()]
         [System.String]
+        $PassiveVoiceEnrollment,
+
+        [Parameter()]
+        [System.String]
         $SpeakerAttributionForBYOD,
 
         [Parameter()]
@@ -75207,6 +75239,10 @@ function Set-CsTeamsCallingPolicy
         [Parameter()]
         [System.String]
         $RealTimeText,
+
+        [Parameter()]
+        [System.String]
+        $ReportCall,
 
         [Parameter()]
         [System.Boolean]
@@ -75482,6 +75518,10 @@ function Set-CsTeamsEventsPolicy
 
         [Parameter()]
         [System.String]
+        $BackroomChat,
+
+        [Parameter()]
+        [System.String]
         $Identity,
 
         [Parameter()]
@@ -75527,6 +75567,10 @@ function Set-CsTeamsEventsPolicy
         [Parameter()]
         [System.String]
         $TownhallEventAttendeeAccess,
+
+        [Parameter()]
+        [System.String]
+        $TownhallMaxResolution,
 
         [Parameter()]
         [System.String]
@@ -75717,6 +75761,10 @@ function Set-CsTeamsMeetingPolicy
         $AllowMeetingCoach,
 
         [Parameter()]
+        [System.Boolean]
+        $AllowMultipleScreenshare,
+
+        [Parameter()]
         [System.String]
         $AttendeeIdentityMasking,
 
@@ -75739,6 +75787,10 @@ function Set-CsTeamsMeetingPolicy
         [Parameter()]
         [System.Boolean]
         $AllowAnonymousUsersToDialOut,
+
+        [Parameter()]
+        [System.String]
+        $BackroomChat,
 
         [Parameter()]
         [System.String]
@@ -75826,7 +75878,15 @@ function Set-CsTeamsMeetingPolicy
 
         [Parameter()]
         [System.String]
+        $PasscodeComplexity,
+
+        [Parameter()]
+        [System.String]
         $ScreenSharingMode,
+
+        [Parameter()]
+        [System.String]
+        $SetRecordingAndTranscriptOwnership,
 
         [Parameter()]
         [System.Boolean]
@@ -78325,6 +78385,14 @@ function New-CsCallQueue
         $CallbackOfferTextToSpeechPrompt,
 
         [Parameter()]
+        [System.String]
+        $TextAnnouncementForCR,
+
+        [Parameter()]
+        [System.String]
+        $TextAnnouncementForCRFailure,
+
+        [Parameter()]
         [System.Int16]
         $TimeoutActionCallPriority,
 
@@ -79205,6 +79273,10 @@ function New-CsTeamsMobilityPolicy
         $IPAudioMobileMode,
 
         [Parameter()]
+        [System.String]
+        $LinksInTeams,
+
+        [Parameter()]
         [System.Management.Automation.SwitchParameter]
         $Confirm,
 
@@ -80026,6 +80098,14 @@ function Set-CsCallQueue
         $CallbackOfferTextToSpeechPrompt,
 
         [Parameter()]
+        [System.String]
+        $TextAnnouncementForCR,
+
+        [Parameter()]
+        [System.String]
+        $TextAnnouncementForCRFailure,
+
+        [Parameter()]
         [System.Int16]
         $TimeoutActionCallPriority,
 
@@ -80783,8 +80863,16 @@ function Set-CsTeamsClientConfiguration
         $ContentPin,
 
         [Parameter()]
+        [System.Boolean]
+        $ExtendedWorkInfoInPeopleSearch,
+
+        [Parameter()]
         [System.String]
-        $MsftInternalProcessingMode
+        $MsftInternalProcessingMode,
+
+        [Parameter()]
+        [System.String]
+        $UseUnifiedDomain
     )
 }
 function Set-CsTeamsCortanaPolicy
@@ -80918,6 +81006,10 @@ function Set-CsTeamsGuestMeetingConfiguration
         [Parameter()]
         [System.Boolean]
         $AllowMeetNow,
+
+        [Parameter()]
+        [System.Boolean]
+        $AllowMultipleScreenshare,
 
         [Parameter()]
         [System.Management.Automation.SwitchParameter]
@@ -81137,6 +81229,10 @@ function Set-CsTeamsMobilityPolicy
         $IPAudioMobileMode,
 
         [Parameter()]
+        [System.String]
+        $LinksInTeams,
+
+        [Parameter()]
         [System.Management.Automation.SwitchParameter]
         $Confirm,
 
@@ -81354,6 +81450,10 @@ function Set-CsTenantFederationConfiguration
         [Parameter()]
         [System.Management.Automation.SwitchParameter]
         $Confirm,
+
+        [Parameter()]
+        [System.String[]]
+        $AllowedTrialTenantDomains,
 
         [Parameter()]
         [System.Boolean]